requirements.txt
flask
python-datauri
flask-cors
